Handle var statement in 3-part for loop
As with the for-in loop, support the syntax where the variable is
declared in the first statement in a 3-statement for loop.

When the vars option is false, this is clearly an error, as it would be
to declare a variable within a block.  However, I contend that when the
vars option is true this should be allowed, as it is legal JavaScript
syntax and is no more confusing than allowing var statements in blocks
(which is allowed without warning when the vars option is true).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>